Teste 2/requirements.txt
pandas == 1.4.2
tabula-py == 2.4.0
PyPDF2 == 2.0.0